# .specify/specs/wechat-login-credit-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: WeChat Login and Credit System API
  description: API for user authentication and credit management system
  version: 1.0.0
  contact:
    name: AI Copywriting Assistant
    url: https://github.com/your-org/copywriting-assistant
servers:
  - url: https://api.copywriting-assistant.com
    description: Production server
  - url: https://staging-api.copywriting-assistant.com
    description: Staging server
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /api/user/profile:
    get:
      summary: Get user profile and credit balance
      description: Retrieves current user profile and credit information
      tags:
        - User
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  credits:
                    $ref: '#/components/schemas/UserCredits'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/credits/deduct:
    post:
      summary: Deduct credits for content generation
      description: Deducts credits from user account for content generation
      tags:
        - Credits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: integer
                  minimum: 1
                  description: Number of credits to deduct
                description:
                  type: string
                  description: Description of the deduction
              required:
                - amount
                - description
      responses:
        '200':
          description: Credits deducted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transaction_id:
                    type: string
                    format: uuid
                  new_balance:
                    type: integer
        '400':
          description: Invalid request or insufficient credits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/credits/history:
    get:
      summary: Get credit transaction history
      description: Retrieves user's credit transaction history
      tags:
        - Credits
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of transactions per page
        - name: type
          in: query
          schema:
            type: string
            enum: [recharge, deduction, bonus, refund]
          description: Filter by transaction type
      responses:
        '200':
          description: Transaction history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/CreditTransaction'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        wechat_openid:
          type: string
        wechat_unionid:
          type: string
          nullable: true
        nickname:
          type: string
        avatar_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UserCredits:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        balance:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreditTransaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        amount:
          type: integer
        transaction_type:
          type: string
          enum: [recharge, deduction, bonus, refund]
        description:
          type: string
        reference_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: supabase-auth-token
...